# Add CUDA depth processor