--- a/.gitignore
+++ b/.gitignore
@@ -2,7 +2,7 @@
 examples/protonect/include/libfreenect2/config.h
 
 # generated resource file
-examples/protonect/src/resources.inc
+examples/protonect/src/resources.inc.h
 examples/protonect/build
 
 # Dependency folders