--- a/examples/protonect/CMakeLists.txt
+++ b/examples/protonect/CMakeLists.txt
@@ -179,7 +179,10 @@ IF(ENABLE_CUDA)
     IF(NOT MSVC)
       SET(CUDA_PIC_FLAGS "-Xcompiler -fPIC")
     ENDIF()
-    CUDA_COMPILE(CUDA_OBJECTS src/cuda_depth_packet_processor.cu OPTIONS ${CUDA_PIC_FLAGS})
+    CUDA_COMPILE(CUDA_OBJECTS
+      src/cuda_depth_packet_processor.cu
+      OPTIONS -use_fast_math ${CUDA_PIC_FLAGS}
+    )
 
     INCLUDE_DIRECTORIES(${CUDA_INCLUDE_DIRS})