--- a/examples/protonect/include/libfreenect2/packet_pipeline.h
+++ b/examples/protonect/include/libfreenect2/packet_pipeline.h
@@ -107,6 +107,17 @@ public:
 };
 #endif // LIBFREENECT2_WITH_OPENCL_SUPPORT
 
+#ifdef LIBFREENECT2_WITH_CUDA_SUPPORT
+class LIBFREENECT2_API CudaPacketPipeline : public BasePacketPipeline
+{
+protected:
+  const int deviceId;
+  virtual DepthPacketProcessor *createDepthPacketProcessor();
+public:
+  CudaPacketPipeline(const int deviceId = -1);
+  virtual ~CudaPacketPipeline();
+};
+#endif // LIBFREENECT2_WITH_CUDA_SUPPORT
 
 } /* namespace libfreenect2 */
 #endif /* PACKET_PIPELINE_H_ */