--- a/examples/protonect/src/packet_pipeline.cpp
+++ b/examples/protonect/src/packet_pipeline.cpp
@@ -136,4 +136,24 @@ DepthPacketProcessor *OpenCLPacketPipeline::createDepthPacketProcessor()
 }
 #endif // LIBFREENECT2_WITH_OPENCL_SUPPORT
 
+#ifdef LIBFREENECT2_WITH_CUDA_SUPPORT
+
+CudaPacketPipeline::CudaPacketPipeline(const int deviceId) : deviceId(deviceId)
+{
+  initialize();
+}
+
+CudaPacketPipeline::~CudaPacketPipeline() { }
+
+DepthPacketProcessor *CudaPacketPipeline::createDepthPacketProcessor()
+{
+  CudaDepthPacketProcessor *depth_processor = new CudaDepthPacketProcessor(deviceId);
+  depth_processor->load11To16LutFromFile("11to16.bin");
+  depth_processor->loadXTableFromFile("xTable.bin");
+  depth_processor->loadZTableFromFile("zTable.bin");
+
+  return depth_processor;
+}
+#endif // LIBFREENECT2_WITH_CUDA_SUPPORT
+
 } /* namespace libfreenect2 */