--- a/examples/protonect/src/resource.cpp
+++ b/examples/protonect/src/resource.cpp
@@ -40,7 +40,7 @@ struct ResourceDescriptor
 };
 
 #ifdef RESOURCES_INC
-#include "resources.inc"
+#include "resources.inc.h"
 #else
 ResourceDescriptor resource_descriptors[] = {};
 #endif