… examples/protonect/include/libfreenect2/config.h.in