Performance:
* GeForce GT 750M - 250Hz
* Tegra K1 - 90Hz

On Tegra K1, numeric optimization improved from 57Hz to 71Hz, and pinned memory improved to 90Hz. Pinned memory required some modifications to `Frame` and `DoubleBuffer`.

This PR needs coordination with #210 because of dependencies.