@larshg 2015-05-26
Tried this on my windows 10 i7 950 with geforce GTX 480 and I get about 2ms processing time ~ 500 hz.

Runs a lot faster than opengl/opencl versions.