@gaborpapp 2015-06-08
I tried building this in OSX, but got linker errors for the cuda object functions:

```
Linking CXX shared library ../lib/libfreenect2.dylib
Undefined symbols for architecture x86_64:
  "std::basic_ios<char, std::char_traits<char> >::widen(char) const", referenced from:
      libfreenect2::CudaDepthPacketProcessorKernelImpl::initDevice(int, unsigned long, unsigned long) in cuda_compile_generated_cuda_depth_packet_processor.cu.o
  "std::ostream::put(char)", referenced from:
      libfreenect2::CudaDepthPacketProcessorKernelImpl::initDevice(int, unsigned long, unsigned long) in cuda_compile_generated_cuda_depth_packet_processor.cu.o
  "std::ostream::flush()", referenced from:
      libfreenect2::CudaDepthPacketProcessorKernelImpl::initDevice(int, unsigned long, unsigned long) in cuda_compile_generated_cuda_depth_packet_processor.cu.o
  "std::ostream& std::ostream::_M_insert<unsigned long>(unsigned long)", referenced from:
      libfreenect2::CudaDepthPacketProcessorKernelImpl::initDevice(int, unsigned long, unsigned long) in cuda_compile_generated_cuda_depth_packet_processor.cu.o
...
```