@xlz 2015-06-08
std::cout here is most innocuous usage. Perhaps there is something wrong in your CUDA toolchain. 

I don't have a Mac with Nvidia card to test. Apparently you are able to build this despite merge conflicts. Maybe you can come up with the solution.