@gaborpapp 2015-06-08
Yes, I had to merge the CMakefile manually.

It seems like a standard c library mixup, libstdc++ vs libc++. I have CUDA 6.5, which seems to require libstdc++, while other libraries are compiled with the default libc++.  I tried to build Protonect with libstdc++ instead. Then the linker finds the missing CUDA functions above, but missing the OpenCV ones for the same reason. Recompiling everything seems like a big hassle.

I am going to try CUDA 7.0 instead, which might work with libc++.