@gaborpapp 2015-06-08
It works with CUDA 7.0. I had to disable OpenCL, otherwise I got duplicate symbol linker error. It might also affect other platforms.

```
Linking CXX shared library ../lib/libfreenect2.dylib
duplicate symbol __ZN12libfreenect223loadBufferFromResourcesERKNSt3__112basic_stringIcNS0_11char_traitsIcEENS0_9allocatorIcEEEEPhm in:
    CMakeFiles/freenect2.dir/src/opencl_depth_packet_processor.cpp.o
    CMakeFiles/freenect2.dir/src/cuda_depth_packet_processor.cpp.o
ld: 1 duplicate symbol for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
```