@xlz 2015-06-08
OK this is a legitimate bug. loadBufferFromResources() is duplicated in both processors. I'll fix it.