@xlz 2015-06-09
What about adding `-Xcompiler -stdlib=libc++` to `CUDA_COMPILE(CUDA_OBJECTS src/cuda_depth_packet_processor.cu OPTIONS ${CUDA_PIC_FLAGS})`?