@xlz 2015-06-09
Work in progress until PR 276, 278 and one more PR finish.