@xlz 2015-06-11
I don't think they are directly comparable. AMD doesn't support CUDA, and Nvidia's OpenCL is obviously worse than CUDA because API abstraction, unoptimized code etc.

I have some math optimization here https://github.com/xlz/libfreenect2/commit/8e5a7c8b3353dd5a3f1446dbe5cacc88504370c3. Most improvement is attributed to `-use_fast_math`, but the OpenCL counterpart of this didn't seem quite effective in my previous testing. There are also some minor math tweaks in this commit that improve things a little bit.

The other optimization is memory access. A CUDA processor with unoptimized memory access works like this: copy the depth packet from the user provided pointer to a page-locked area, send it to
GPU, and finally send the result to CPU.

* The first copy can be removed by making the memory of the depth packet page-locked in the first place with CUDA's "pinned memory" https://github.com/xlz/libfreenect2/commit/1024bc87b0040ad12323af188985d97db0b28523. OpenCL has equivalent for this. This requires modifications to `Frame` and `DoubleBuffer` to allow custom GPU-aware memory allocator.
* The copy from CPU to GPU can be avoided with mapped memory, but only on integrated platforms where CPU and GPU share the same memory. PCIe bus transmission has to happen if GPU doesn't share memory with CPU. However mapping memory on Tegra seems to reduce the performance of its JPEG decoder so I didn't post that commit here.
* Finally, the result on GPU doesn't need to be sent back to CPU at all if it is only for displaying or further computation on the same GPU.

float3 is internally float4 so I don't think that matters.