@xlz 2015-06-26
Current status of this PR: The CUDA runtime setup, CUDA kernels, and memory management are complete. It is broken by recent changes in the CMake build system (change of context, addition of static/shared build configurations). This can be fixed but I don't have a Kinect 2 to test recently, and anyone is welcome to work upon the current code.

I am still uncertain about how to best interface with other parts involving minimal memory copy. In this PR, https://github.com/xlz/libfreenect2/commit/de911375b4d63dc7c909cc6d622ee2c1c5db2087 `Allow processors to define custom Frame classes` and passes CUDA "pinned memory" to the frame listener to avoid a memory copy internal in CUDA. It might be better to implement zero-copy between the depth processor and the frame listener `listener.waitForNewFrame()` by passing a pointer to a fixed buffer defined by the processor. This way avoids allocating, page locking, and freeing 848KB 30 times per second. The problems with this:
  - It requires a double buffer structure between the depth processor and the frame listener, and changes to Protonect.cpp and iai_kinect2.
  - Coordination with OpenCL processor. The OpenCL processor currently does not use zero-copy so this interface change might also impact it.
  - Coordination with VAAPI color processor.
  - Coordination with Tegra color processor.
  - In my previous testing on Jetson, zero-copy did not seem vastly better than one-copy, and zero-copy seemed to interfere with Tegra libjpeg. This might be different on other platforms.
  - Optionally avoiding GPU to CPU copy inside the depth processor if the output of the depth processor is for display or further GPU-only computation. (Intel: OpenCL + VA-API, Nvidia: CUDA + VDPAU/VA-API)

To help with understanding the data flow, it can be illustrated as:

```
libusb -> stream parsers -> packet processors -> frame listeners
```