@wiedemeyer 2016-02-16
@xlz
True, one could compute those values for the complete p0 table.
I created a float3 buffer with 512 * 424 * 6 entries (cos and sin for each of the 3 phases). But it seems that computing it on the GPU is actually faster that the lookup. The time increased from around 1.7 ms to 1.9 ms. Did you tried it on the CUDA implementation?
Other improvements are to use `sincos` to compute both at once or `native_sin` and `native_cos`. I think the last option is the fastest, but both were around 1.7 to 1.75 ms, so no significant difference. Maybe I can check how it differs on slower GPUs later.
By the way, the CUDA processor is running smoothly at 1 ms, but really great is the VAAPI jpeg decoding, 5 ms instead of 15 ms. That is an awesome improvement, great job!